+++ plone/app/users/tests/base.py
@@ -6,12 +6,10 @@ without the PloneTestCase.setupPloneSite() side effects.
 """
 
 from Products.PloneTestCase.PloneTestCase import FunctionalTestCase
-
 from Acquisition import aq_base
 from zope.component import getSiteManager
 from Products.CMFPlone.tests.utils import MockMailHost
 from Products.MailHost.interfaces import IMailHost
-from Products.CMFCore.utils import getToolByName
 
 # BBB Zope 2.12
 try:
@@ -21,20 +19,19 @@ except ImportError:
 
 from AccessControl.SecurityInfo import ClassSecurityInfo
 from OFS.Cache import Cacheable
-from Products.PluggableAuthService.interfaces.plugins import IValidationPlugin, IPropertiesPlugin
+from Products.PluggableAuthService.interfaces.plugins import IValidationPlugin
 from Products.PluggableAuthService.plugins.BasePlugin import BasePlugin
 from Products.PluggableAuthService.utils import classImplements
 from Products.CMFCore.interfaces import ISiteRoot
-from zope.component import getUtility, getAdapter
+from zope.component import getUtility
 from Products.PlonePAS.Extensions.Install import activatePluginInterfaces
 
 
-class TestCase(FunctionalTestCase):
+class BaseTestCase(FunctionalTestCase):
     """base test case which adds amin user"""
 
-
     def afterSetUp(self):
-        super(TestCase, self).afterSetUp()
+        super(BaseTestCase, self).afterSetUp()
         self.browser = Browser()
         self.portal.acl_users._doAddUser('admin', 'secret', ['Manager'], [])
 
@@ -64,27 +61,26 @@ class TestCase(FunctionalTestCase):
 
     def activateDefaultPasswordPolicy(self):
         uf = self.portal.acl_users
-        plugins = uf._getOb('plugins')
         for policy in uf.objectIds(['Default Plone Password Policy']):
             activatePluginInterfaces(self.portal, policy)
-            validators = plugins.listPlugins(IValidationPlugin)
-            #assert policy in validators
 
     def beforeTearDown(self):
         self.portal.MailHost = self.portal._original_MailHost
         sm = getSiteManager(context=self.portal)
         sm.unregisterUtility(provided=IMailHost)
-        sm.registerUtility(aq_base(self.portal._original_MailHost), provided=IMailHost)
+        sm.registerUtility(
+            aq_base(self.portal._original_MailHost),
+            provided=IMailHost
+        )
 
         portal = getUtility(ISiteRoot)
         pas_instance = portal.acl_users
-        plugin = getattr(pas_instance,'test', None)
+        plugin = getattr(pas_instance, 'test', None)
         if plugin is not None:
             plugins = pas_instance._getOb('plugins')
             plugins.deactivatePlugin(IValidationPlugin, 'test')
             #plugins.deactivatePlugin(IPropertiesPlugin, 'test')
             pas_instance.manage_delObjects('test')
-
 
     def setMailHost(self):
         self.portal.MailHost.smtp_host = 'localhost'
@@ -94,8 +90,11 @@ class TestCase(FunctionalTestCase):
         self.portal.MailHost.smtp_host = ''
         setattr(self.portal, 'email_from_address', '')
 
-# Dummy password validation PAS plugin
+    def test_nothing(self):
+        """Add a dummy test here, so the base class 'passes'."""
+        pass
 
+# Dummy password validation PAS plugin
 
 
 class DeadParrotPassword(BasePlugin, Cacheable):
@@ -107,18 +106,16 @@ class DeadParrotPassword(BasePlugin, Cacheable):
         self.title = title
 
     security.declarePrivate('validateUserInfo')
-    def validateUserInfo(self, user, set_id, set_info ):
 
+    def validateUserInfo(self, user, set_id, set_info):
         errors = []
-
         if set_info and set_info.get('password', None) is not None:
             password = set_info['password']
             if password.count('dead') or password == '':
-                errors = [{'id':'password','error':u'Must not be dead'}]
+                errors = [{'id': 'password', 'error': u'Must not be dead'}]
             else:
                 errors = []
         return errors
 
 
-classImplements(DeadParrotPassword,
-                IValidationPlugin)
+classImplements(DeadParrotPassword, IValidationPlugin)